Added files to interface to bigquery from older project
I added files that contains lines of code that allows to interface with bigquery trough the python API from an older project, namely the one from https://github.com/davide-chiuchiu/BigQuery_on_Stackoverflow_data.

The interface functions do not reflect the specific structure of this project, and needs to be updated accordingly.