Project pivot and completed exploratory analysis
Since last time, I realized that the project for the google analytics was not particularly interesting because the dataset itself was poorly documented. As a consequence, I decided to reinvent the project and develop a simple recommendation engine of musical groups using implicit listening data from the listenbrainz dataset in bigquery. The pivot required me to change a few of the sql queries I already wrote, but also to redo the exploratory analysis. This commit contains all these new parts of the project.