Completed data preprocessing
I completed the data preprocessing phase before starting with the keras model itself. The preprocessing assumes that keras will see in input data from a pandas dataframe, and I will stick with this simple example for the moment. Later iterations of this project will try to remove this intermediate step and feed the data directly from bigquery.

In addition, this commit contains a new file, preprocessing_utilities.py where I put some preprocessing utilities that I use, but that are not particularly relevant for the project scope itself.