Implemented functions that allows to query a recommendation
I have updated the main python notebook with a collection of function that uses the tensorflow model implemented in the previous commits to recommend new artists to a user. Specifically, the functions targets artists that are not in the user library, and the output is already in a human readable format.