Added regularization to the embedding layers to reduce anomalous results
I updated the model so that I now added regularizations in the embeddings. This introduce a step in which a sparse layer is converted into a dense one, which is far from optimal, but it reduces anomalous recommendations.

In this commit I also tried to account for non-repetible results, but this specific problem uses stochastic optimization, which is non-reproducible in most cases. Hence, yay!